pyproject.toml
[project]
name = "medical-extraction"
version = "0.1.0"
description = "Medical data extraction with BAML and DGraph integration"
authors = [
    {name = "Medical Extraction Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "baml-py>=0.60.0",
    "pydgraph>=23.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[tool.uv]
dependencies = [
    "baml-py>=0.60.0",
    "pydgraph>=23.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
